requirements.txt
beautifulsoup4>=4.10.0
python-dotenv>=0.19.2
pyTelegramBotAPI>=4.2.0
python-dateutil>=2.8.1
